python/pyproject.toml
[build-system]
requires = ["setuptools>=68,<75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "byzpy"
version = "0.1.0"
description = "A research framework for Byzantine-robust distributed ML"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "ByzPy contributors", email = "byzpy@users.noreply.github.com" }]
maintainers = [{ name = "ByzPy contributors", email = "byzpy@users.noreply.github.com" }]
license = { text = "MIT" }
dependencies = [
    "torch>=2.2,<2.6",
    "torchvision>=0.17,<0.21",
    "numpy>=1.24,<2.1",
    "matplotlib>=3.8,<3.11",
    "cloudpickle>=2.2,<4.0",
]
keywords = ["distributed-learning", "byzantine", "federated-learning", "pytorch", "robust-ml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/Byzpy/byzpy"
Documentation = "https://github.com/Byzpy/byzpy/tree/main/docs"
Source = "https://github.com/Byzpy/byzpy"
Issues = "https://github.com/Byzpy/byzpy/issues"

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x>=13.2.0",
    "ucxx-cu12>=0.38",
]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.23.7",
    "pytest-cov==5.0.0",
]

[project.scripts]
byzpy = "byzpy.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["byzpy*"]

[tool.pytest.ini_options]
markers = [
    "real_actor_backends: run tests against the real actor backends instead of stubs",
]
